model/etl/requirements.txt
elasticsearch===8.5.1
pandas
openpyxl
numpy
python-dateutil
python-dotenv